+++ app/view/positioner.cpp
@@ -931,7 +931,7 @@ void Positioner::initSignalingForLocationChangeSliding()
 
             //! make sure that View has been repositioned properly in next screen edge and show view afterwards
             if (isrelocationlastevent) {
-                QTimer::singleShot(100, [this]() {
+                QTimer::singleShot(100, this, [this]() {
                     onLastRepositionApplyEvent();
                 });
             }
@@ -958,7 +958,7 @@ void Positioner::initSignalingForLocationChangeSliding()
 
             //! make sure that View has been repositioned properly in next screen and show view afterwards
             if (isrelocationlastevent) {
-                QTimer::singleShot(100, [this]() {
+                QTimer::singleShot(100, this, [this]() {
                     onLastRepositionApplyEvent();
                 });
             }
@@ -973,7 +973,7 @@ void Positioner::initSignalingForLocationChangeSliding()
 
             //! make sure that View has been repositioned properly in next layout and show view afterwards
             if (isrelocationlastevent) {
-                QTimer::singleShot(100, [this]() {
+                QTimer::singleShot(100, this, [this]() {
                     onLastRepositionApplyEvent();
                 });
             }